Explicitly version numpy in recipes
Otherwise, the automatic version resolution algorithm in conda may sometimes
pick up different versions of numpy (depending on dependencies of a
particular package).  E.g., if a package does not depend on a package that
depends on numpy, it will pick the latest version (e.g., 1.10), but if it
does (e.g., matplotlib right now depends on 1.9), it will pick an older
version leading to binary incompatibility.