Generate individual metapkgs w. EUPS files for pkgs provided by conda
Instead of having all table files (and other ups/* data) packed in a single
package, generate a conda package with the relevant EUPS file for each
internal package.  They will be named lsst-PACKAGE-eups-configs and contain
the contents of the ups/ directory (and be declared as an EUPS product).

Their generation differs from generating other packages only in that they
use a special build.sh template (build-internal.sh.template), which copies
the ups/* directory instead of doing a full build.  This simplifies the
internal code logic.

This change makes the lsst-product-configs unnecessary.