bin/make-patch: a utility for easy creation/maintenence of patches
To make it easy to create a patch from a brancg in a git repo, use
the make-patch utility as:

    make-patch patch-name.patch master conda-patches

This creates a file patch-name.patch with a diff between
master..conda-patches in the etc/patches/<product>/ directory, where
<product> is automatically inferred from the name of the .table file in
the git repo's ups/ directory. If you omit the third argument to
make-patch, HEAD is assumed. If you omit the second argument, master
is assumed. Therefore, assuming you're on conda-patches branch, running:

    make-patch patch-name.patch

is equivalent to the invocation above.